Create UTM-TRACER-WOOCOMMERCE
I have developed an additional functionality for WooCommerce that helps in case you have different organised campaigns with UTM tags. The purpose of the development is to track the cookie that is formed by the free version of Handle Utm Grabber and fill the data in a small table in the order overview section. Thus, the order can be tracked whether it comes from an advertising campaign or from another source of information.